requirements.txt
absl-py==0.11.0
aiohttp==3.6.2
alabaster==0.7.12
alembic==1.4.2
appdirs==1.4.4
appnope==0.1.0
argon2-cffi==20.1.0
arrow==0.15.7
astroid==2.4.2
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==19.3.0
aws-lambda-builders==0.6.0
aws-sam-cli==0.33.1
aws-sam-translator==1.35.0
aws-xray-sdk==2.6.0
Babel==2.8.0
backcall==0.2.0
binaryornot==0.4.4
black==19.10b0
bleach==3.2.1
blis==0.7.4
boto==2.49.0
boto3==1.14.18
botocore==1.17.18
cached-property==1.5.2
cachetools==4.2.0
catalogue==2.0.6
Cerberus==1.3.2
certifi==2020.6.20
cffi==1.14.0
cfn-lint==0.33.2
chardet==3.0.4
chevron==0.13.1
click==7.1.2
codecov==2.1.7
commonmark==0.9.1
configparser==5.0.0
contextlib2==0.6.0.post1
cookiecutter==1.6.0
coverage==5.2
cryptography==2.9.2
cycler==0.10.0
cymem==2.0.5
dateparser==0.7.6
decorator==4.4.2
deep-daze==0.8.0
deepmerge==0.2.1
defusedxml==0.6.0
dependency-injector==4.31.1
Deprecated==1.2.13
distlib==0.3.1
docker==4.2.2
docutils==0.15.2
ecdsa==0.14
einops==0.3.0
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.1.0/en_core_web_md-3.1.0-py3-none-any.whl
entrypoints==0.3
filelock==3.0.12
fire==0.4.0
flake8==3.8.3
Flask==1.1.2
flatbuffers==1.12
ftfy==5.9
future==0.18.2
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
humanfriendly==8.2
idna==2.8
imageio==2.9.0
imagesize==1.2.0
importlib-metadata==1.7.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jaeger-client==4.4.0
jedi==0.17.2
Jinja2==2.11.2
jinja2-time==0.2.0
jmespath==0.10.0
joblib==0.16.0
jsondiff==1.1.2
jsonpatch==1.26
jsonpickle==1.4.1
jsonpointer==2.0
jsonschema==3.2.0
junit-xml==1.9
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.1
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
Mako==1.1.3
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
mistune==0.8.4
mock==4.0.2
more-itertools==8.4.0
moto==1.3.14
mpi4py==3.0.3
multidict==4.7.6
murmurhash==1.0.5
nbclient==0.5.0
nbconvert==6.0.3
nbformat==5.0.7
nest-asyncio==1.4.0
networkx==2.4
nltk==3.5
notebook==6.1.4
numpy==1.19.4
oauthlib==3.1.0
opentracing==2.4.0
opt-einsum==3.3.0
packaging==20.4
pandas==0.23.4
pandocfilters==1.4.2
parso==0.7.1
pathspec==0.8.0
pathy==0.6.0
pbr==5.4.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pipenv==2020.11.15
plotly==4.1.0
pluggy==0.13.1
ply==3.11
poyo==0.5.0
preshed==3.0.5
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.12.2
psutil==5.7.0
psycopg2==2.9.1
ptyprocess==0.6.0
py==1.9.0
py-zipkin==0.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.8.2
pyenchant==3.2.0
pyflakes==2.2.0
PyGithub==1.55
Pygments==2.6.1
PyJWT==2.1.0
pylint==2.5.3
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.16.0
pysummarization==1.1.7
pytest==5.4.3
pytest-asyncio==0.14.0
pytest-cov==2.10.0
python-dateutil==2.8.0
python-dotenv==0.15.0
python-editor==1.0.4
python-jose==3.1.0
python-json-logger==0.1.11
pytorch-ranger==0.1.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
recommonmark==0.6.0
regex==2020.6.8
requests==2.24.0
requests-oauthlib==1.3.0
responses==0.10.15
retrying==1.3.3
rsa==4.6
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
sacremoses==0.0.43
schema==0.7.4
scikit-learn==0.23.1
scipy==1.5.1
Send2Trash==1.5.0
serverlessrepo==0.1.9
siren-pytorch==0.0.9
six==1.15.0
smart-open==5.2.1
snowballstemmer==2.0.0
spacy==3.1.2
spacy-legacy==3.0.8
Sphinx==3.1.2
sphinx-click==2.3.2
sphinx-rtd-theme==0.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-spelling==7.1.0
SQLAlchemy==1.3.18
SQLAlchemy-Utils==0.36.7
srsly==2.4.1
sshpubkeys==3.1.0
tabulate==0.8.7
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
tensorflow-hub==0.9.0
tensorflowjs==2.8.2
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
thinc==8.0.10
threadloop==1.0.2
threadpoolctl==2.1.0
thrift==0.13.0
thriftpy2==0.4.11
tokenizers==0.10.2
toml==0.10.1
tomlkit==0.5.8
torch==1.7.1
torch-optimizer==0.1.0
torchaudio==0.7.2
torchvision==0.8.2
tornado==6.0.4
tqdm==4.55.1
traitlets==5.0.4
transformers==4.5.0
typed-ast==1.4.1
typer==0.3.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.9
virtualenv==20.3.1
virtualenv-clone==0.5.4
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
whichcraft==0.6.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xmltodict==0.12.0
yarl==1.4.2
zipp==3.1.0
